requirements.txt
pydantic
google-genai
python-dotenv
litellm
evidently